OneDrive/Desktop/Nexora-main 2/Nexora-main/ai-service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
anyio
